I DID IT. THE VIDEO SYNC IS WORKINGGGG CORRECTLY. FINALLY, YESS!!
Fixed the isPaused function to be async.

Fixed the time conversion which was syncing the video to a large number which basically ended the video on trying to sync.

Removed formatTime function which was causing issues exchanging commands (could have had this optionally, but I am not sure why it wasn't working. Removed as it wasn't necessary for the functioning, only for debugging).

Haven't yet removed debug console logs at every point. TBD if to be kept.